Refactor location of tests
Organize file structure following Pytest best practices, and separate
tests into different files.
