+++ tox.ini
@@ -8,7 +8,7 @@ passenv = TERM
 commands =
     coverage erase
     pytest --cov=tokendito --cov-append -v -ra -k 'unit' -s tests/ --
-    pytest --cov=tokendito --cov-append -v -ra -k 'functional and not credentials' -s tests/ --
+    pytest --cov=tokendito --cov-append -v -ra -k 'functional and not auth' -s tests/ --
 
 [testenv:lint]
 skip_install = true
@@ -20,7 +20,7 @@ commands =
 skip_install = true
 passenv = TOKENDITO_*
 commands =
-    pytest --cov=tokendito --cov-append -v -rA -k 'functional and (config or credentials)' -s tests/ {posargs}
+    pytest --cov=tokendito --cov-append -v -rA -k 'functional and auth' -s tests/ {posargs}
 
 [testenv:coverage]
 skip_install = true
